// src/api-client/models/PlannedExpensesResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Budget Manager API
 * Personal finance management API for tracking income, expenses, accounts, and budgets.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PlannedExpenseViewDto } from './PlannedExpenseViewDto';
import {
    PlannedExpenseViewDtoFromJSON,
    PlannedExpenseViewDtoFromJSONTyped,
    PlannedExpenseViewDtoToJSON,
    PlannedExpenseViewDtoToJSONTyped,
} from './PlannedExpenseViewDto';
import type { BudgetSectionItemDto } from './BudgetSectionItemDto';
import {
    BudgetSectionItemDtoFromJSON,
    BudgetSectionItemDtoFromJSONTyped,
    BudgetSectionItemDtoToJSON,
    BudgetSectionItemDtoToJSONTyped,
} from './BudgetSectionItemDto';

/**
 * 
 * @export
 * @interface PlannedExpensesResponseDto
 */
export interface PlannedExpensesResponseDto {
    /**
     * 
     * @type {number}
     * @memberof PlannedExpensesResponseDto
     */
    total?: number;
    /**
     * 
     * @type {Array<PlannedExpenseViewDto>}
     * @memberof PlannedExpensesResponseDto
     */
    plannedExpenses?: Array<PlannedExpenseViewDto> | null;
    /**
     * 
     * @type {Array<BudgetSectionItemDto>}
     * @memberof PlannedExpensesResponseDto
     */
    items?: Array<BudgetSectionItemDto> | null;
}

/**
 * Check if a given object implements the PlannedExpensesResponseDto interface.
 */
export function instanceOfPlannedExpensesResponseDto(value: object): value is PlannedExpensesResponseDto {
    return true;
}

export function PlannedExpensesResponseDtoFromJSON(json: any): PlannedExpensesResponseDto {
    return PlannedExpensesResponseDtoFromJSONTyped(json, false);
}

export function PlannedExpensesResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlannedExpensesResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'total': json['total'] == null ? undefined : json['total'],
        'plannedExpenses': json['plannedExpenses'] == null ? undefined : ((json['plannedExpenses'] as Array<any>).map(PlannedExpenseViewDtoFromJSON)),
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(BudgetSectionItemDtoFromJSON)),
    };
}

export function PlannedExpensesResponseDtoToJSON(json: any): PlannedExpensesResponseDto {
    return PlannedExpensesResponseDtoToJSONTyped(json, false);
}

export function PlannedExpensesResponseDtoToJSONTyped(value?: PlannedExpensesResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'total': value['total'],
        'plannedExpenses': value['plannedExpenses'] == null ? undefined : ((value['plannedExpenses'] as Array<any>).map(PlannedExpenseViewDtoToJSON)),
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(BudgetSectionItemDtoToJSON)),
    };
}